REVIEW: work at category logic

+++ src/main/java/ru/ai/sin/service/impl/CategoryService.java
@@ -12,7 +12,7 @@ public interface CategoryService {
     // ---------- GET METHODS ----------
     CategoryDTO getById(Long id);
 
-    List<CategoryDTO> getAll(long page, long size);
+    List<CategoryDTO> getAll(int page, int size);
 
     // ---------- POST METHODS ----------
     CategoryDTO create(AddCategoryReq addCategoryReq);
